requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.7
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-docx==0.8.11
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.14
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
